Number of variables is unsigned that is return by all SAT solvers
No variables is 0. More than one is positive. Negative makes no sense
and no SAT solver returns that number that I know of.

Some SAT solvers (maybe newer ones) might need more than just unsigned
for the number of variables. So I'm making the return value unsigned
long instead of simply unsigned.

+++ src/sat/CryptoMinisat.cpp
@@ -66,7 +66,7 @@ namespace BEEV
     s->verbosity = v;
   }
 
-  int CryptoMinisat::nVars()
+  unsigned long CryptoMinisat::nVars()
   {return s->nVars();}
 
   void CryptoMinisat::printStats()
