+++ src/sat/MinisatCore.h
@@ -49,7 +49,7 @@ namespace BEEV
 
     void setVerbosity(int v);
 
-    int nVars();
+    unsigned long nVars();
 
     void printStats();
